Adds IE CORS workaround to all the example pages so they work correctly when loaded from a file:// URL.
To view the examples on IE, right-click on the index.html file and any
example HTML file that you wish to use and select "Properties". From
this dialog click on the "Unblock" button in the security section.
When you then load the page in IE (by double clicking on it) a popdown
will appear telling you that ActiveX controls are restricted, click on
it and select "Allow Blocked Content...", this will allow the page to
make an outbound CORS request to the engine.

If you don't follow these instructions on IE, then you'll get a
"Permission Denied" error in the search response JSON object.